+++ baybe/surrogates/custom.py
@@ -10,11 +10,11 @@ It is planned to solve this issue in the future.
 
 from __future__ import annotations
 
-from collections.abc import Callable
-from typing import TYPE_CHECKING, Any, ClassVar
+from typing import TYPE_CHECKING, ClassVar, NoReturn
 
 from attrs import define, field, validators
 
+from baybe.exceptions import DeprecationError
 from baybe.parameters import (
     CategoricalEncoding,
     CategoricalParameter,
@@ -24,10 +24,8 @@ from baybe.parameters import (
     TaskParameter,
 )
 from baybe.searchspace import SearchSpace
-from baybe.serialization.core import block_serialization_hook, converter
-from baybe.surrogates.base import GaussianSurrogate, Surrogate
-from baybe.surrogates.utils import batchify, catch_constant_targets
-from baybe.surrogates.validation import validate_custom_architecture_cls
+from baybe.surrogates.base import GaussianSurrogate
+from baybe.surrogates.utils import batchify
 from baybe.utils.numerical import DTypeFloatONNX
 
 if TYPE_CHECKING:
@@ -35,80 +33,12 @@ if TYPE_CHECKING:
     from torch import Tensor
 
 
-def register_custom_architecture(
-    joint_posterior_attr: bool = False,
-    constant_target_catching: bool = True,
-    batchify_posterior: bool = True,
-) -> Callable:
-    """Wrap a given custom model architecture class into a ```Surrogate```.
-
-    Args:
-        joint_posterior_attr: Boolean indicating if the model returns a posterior
-            distribution jointly across candidates or on individual points.
-        constant_target_catching: Boolean indicating if the model cannot handle
-            constant target values and needs the @catch_constant_targets decorator.
-        batchify_posterior: Boolean indicating if the model is incompatible
-            with t- and q-batching and needs the @batchify decorator for its posterior.
-
-    Returns:
-        A function that wraps around a model class based on the specifications.
-    """
-
-    def construct_custom_architecture(model_cls):
-        """Construct a surrogate class wrapped around the custom class."""
-        validate_custom_architecture_cls(model_cls)
-
-        class CustomArchitectureSurrogate(Surrogate):
-            """Wraps around a custom architecture class."""
-
-            joint_posterior: ClassVar[bool] = joint_posterior_attr
-            supports_transfer_learning: ClassVar[bool] = False
-
-            def __init__(self, *args, **kwargs):
-                self._model = model_cls(*args, **kwargs)
-
-            def _fit(self, train_x: Tensor, train_y: Tensor, context: Any) -> None:
-                return self._model._fit(train_x, train_y, context)
-
-            def _posterior(self, candidates: Tensor, /) -> tuple[Tensor, Tensor]:
-                return self._model._posterior(candidates)
-
-            def __get_attribute__(self, attr):
-                """Access the attributes of the class instance if available.
-
-                If the attributes are not available,
-                it uses the attributes of the internal model instance.
-                """
-                # Try to retrieve the attribute in the class
-                try:
-                    val = super().__getattribute__(attr)
-                except AttributeError:
-                    pass
-                else:
-                    return val
-
-                # If the attribute is not overwritten, use that of the internal model
-                return self._model.__getattribute__(attr)
-
-        # Catch constant targets if needed
-        cls = (
-            catch_constant_targets(CustomArchitectureSurrogate)
-            if constant_target_catching
-            else CustomArchitectureSurrogate
-        )
-
-        # batchify posterior if needed
-        if batchify_posterior:
-            cls._posterior = batchify(cls._posterior)
-
-        # Block serialization of custom architectures
-        converter.register_unstructure_hook(
-            CustomArchitectureSurrogate, block_serialization_hook
-        )
-
-        return cls
-
-    return construct_custom_architecture
+def register_custom_architecture(*args, **kwargs) -> NoReturn:
+    """Deprecated! Raises an error when used."""  # noqa: D401
+    raise DeprecationError(
+        "The 'register_custom_architecture' decorator is no longer available. "
+        "There will be a replacement based on Python's `typing.Protocol` soon."
+    )
 
 
 @define(kw_only=True)